src/test/java/by/itacademy/reflection/generator/FieldGeneratorTest.java
package by.itacademy.reflection.generator;

import by.itacademy.reflection.Student;
import org.junit.Assert;
import org.junit.Test;

public class FieldGeneratorTest {
    @Test
    public void studentFieldsGenerationValid() throws CustomInstantiationException {
        BeanFactory<Student> studentBeanFactory = new BeanFactory<>(Student.class);
        Student student = studentBeanFactory.createObject();
        FieldGenerator<Student> fieldGenerator = new FieldGenerator<>(student);
        fieldGenerator.execute();

        Assert.assertNotNull(student);
        Assert.assertNotNull(student.getName());
        Assert.assertEquals(33, student.getAge());
    }
}
